model_template/examples/sam3/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pyyaml>=6.0.0
pillow>=10.0.0
numpy>=1.26.0
# Uncomment when using real SAM:
# sam2 @ git+https://github.com/facebookresearch/sam2.git
